refactor: Relocate enums and adjust rental status logic (#181)
- Move `RentingStatusModel` to the `common.enums` package and rename it to `RentingStatus` for better code organization.
- Update all import paths for the moved enum.
- Correct the return date calculation in `RentalDetailStatusMapper` by removing an off-by-one error.
- Change the color for the `RETURNED` status to `Gray400`.
- Delete the unused `ResvStatus` enum file.
- Remove `@Singleton` from the `HttpLoggingInterceptor` provider in `NetworkModule`.
