+++ app/src/main/java/com/example/rentit/domain/rental/model/RentalDetailStatusModel.kt
@@ -1,6 +1,7 @@
 package com.example.rentit.domain.rental.model
 
 import com.example.rentit.common.enums.RentalStatus
+import com.example.rentit.common.enums.RentingStatus
 import com.example.rentit.common.uimodel.RentalSummaryUiModel
